+++ lib/flutter_instagram_image_picker.dart
@@ -1,5 +1,7 @@
 library flutter_instagram_image_picker;
 
-export 'picker.dart';
-export 'model/photo.dart';
-export 'instagram_auth.dart';
+/// A Calculator.
+class Calculator {
+  /// Returns [value] plus 1.
+  int addOne(int value) => value + 1;
+}