+++ lib/ui/photo_grid.dart
@@ -1,49 +1,50 @@
 import 'package:flutter/material.dart';
+import '../model/photo.dart';
 
 import '../debounce.dart';
-import '../flutter_instagram_image_picker.dart';
 
 /// Widget that shows list of photos in a grid
 class PhotoGrid extends StatefulWidget {
   final List<Photo> _photos;
-  final List<Photo> _selectedPhotos;
+  final List<Photo>? _selectedPhotos;
   final Function(Photo) onPhotoTap;
   final Function onLoadMore;
 
-  PhotoGrid(
+  const PhotoGrid(
     this._photos,
     this._selectedPhotos, {
-    @required this.onPhotoTap,
-    @required this.onLoadMore,
-  });
+    Key? key,
+    required this.onPhotoTap,
+    required this.onLoadMore,
+  }) : super(key: key);
 
   @override
   PhotoGridState createState() {
-    return new PhotoGridState();
+    return PhotoGridState();
   }
 }
 
 class PhotoGridState extends State<PhotoGrid> {
-  ScrollController _controller;
-  Debouncer _debouncer;
+  ScrollController? _controller;
+  Debouncer? _debouncer;
 
   @override
   void initState() {
     super.initState();
     _debouncer = Debouncer(
-      Duration(seconds: 1),
+      const Duration(seconds: 1),
       (arg) => widget.onLoadMore(),
       null,
       true,
     );
 
     _controller = ScrollController();
-    _controller.addListener(() {
-      double maxScroll = _controller.position.maxScrollExtent;
-      double currentScroll = _controller.position.pixels;
+    _controller!.addListener(() {
+      double maxScroll = _controller!.position.maxScrollExtent;
+      double currentScroll = _controller!.position.pixels;
       double delta = 100.0;
       if (maxScroll - currentScroll <= delta) {
-        _debouncer.debounce();
+        _debouncer?.debounce();
       }
     });
   }
@@ -51,11 +52,11 @@ class PhotoGridState extends State<PhotoGrid> {
   @override
   void dispose() {
     super.dispose();
-    _controller.dispose();
+    _controller?.dispose();
   }
 
   bool _isSelected(Photo photo) {
-    return widget._selectedPhotos.indexOf(photo) != -1;
+    return widget._selectedPhotos?.contains(photo) ?? false;
   }
 
   Widget _buildPhotoCell(Photo photo) {
@@ -66,14 +67,14 @@ class PhotoGridState extends State<PhotoGrid> {
           Positioned.fill(
             child: FadeInImage(
               image: NetworkImage(photo.url),
-              placeholder: AssetImage('assets/loading.gif',
+              placeholder: const AssetImage('assets/loading.gif',
                   package: 'flutter_instagram_image_picker'),
               fit: BoxFit.cover,
             ),
           ),
           if (_isSelected(photo))
             Opacity(opacity: 0.5, child: Container(color: Colors.blue))
-        ].where((o) => o != null).toList(),
+        ].toList(),
       ),
     );
   }
@@ -88,7 +89,7 @@ class PhotoGridState extends State<PhotoGrid> {
         controller: _controller,
         children: List.generate(widget._photos.length, (index) {
           return Padding(
-            padding: EdgeInsets.all(8.00),
+            padding: const EdgeInsets.all(8.00),
             child: _buildPhotoCell(widget._photos[index]),
           );
         }),